[AAE-37713] moved getLocale to translation service
