+++ lib/core/src/lib/mock/translation.service.mock.ts
@@ -47,6 +47,8 @@ export class TranslationMock implements TranslationService {
         return of(key);
     }
 
+    getLocale(): any {}
+
     instant(key: string | Array<string>): string | any {
         return key;
     }
